Add scoring and restructure game
Create a game helper function ouside the game function to run the game
and stop when a win or lose condition is met.  Add scoring to the game
based on the traditional sports point system where a tie is a half of a
win and loss.  Remove the event listener when at least 5 wins or losses
occur and display final game result.